requirements.txt
aiogram>=3.16.0
deepgram-sdk>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
loguru>=0.7.0
anthropic>=0.18.1
certifi>=2023.11.17
